requirements.txt
streamlit
pandas
matplotlib
numpy
xlrd
opencv-python
pillow
datetime
folium
streamlit-folium
